requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
insightface==0.7.3
onnxruntime-gpu==1.16.3
opencv-python==4.9.0.80
numpy==1.24.3
Pillow==10.1.0
pydantic==2.5.3
